[AF041]Basic Activty image views and discriptions
Deleted the empty MainActivty and Layout.
Created a new one from the Basic Activity template.
In the content_main.xml I added three image views with content description and saved that description in res > strings.xml. The same content description is used for the text of the text view next to image view.
The option "Match constrain" is used for the text view to match the content to the constrains.

In Main Activity I added a onClick method to resolve the click of any image view thaht has the onClick attribute with the name of the method. The method has to be public void and has to take a View obj as argument.

The takeOrder method then shows a Toast with a message that is initialisied wit a string res for every particular image view, that was clicked.