[AF041]Start Act2 with Image Click and show order
Changed the icon of th fab to a new icon i created in the drawable folder.
The fab start the Act2 with an intent but without any inputs.
The takeOrder method now calls the startAct2WithOrder method that started a new Act2 with an intent and String Extra that hold the right order for the choosen image.

In Act2 that order message is loaded to a textview and the value of it is saved in a bundle in onSaveInstanceState when the activity is reloaded for example when the oriantaion is changed.