[AF042] Input controls Codelabs
Deleted not needet fragments.
In Act2 I created multiple edit text views with different input attributes like for example  phone for the number.
I also created a RadioGroup with three RadioButton an a onClick method that responds to the right radio button.
The Radiobutton rb_nextday is initialized  to state chacked when the Act2 is created.